requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
pyyaml==5.3.1
snowflake-connector-python[pandas]==2.3.3
pandas==1.1.3
numpy==1.19.3
jinjasql==0.1.8